+++ docs/CHANGELOG.md
@@ -6,6 +6,7 @@
 - `invite` command.
 - Error when giveaway result is too big is handled.
 - `avatar`'s embed description simplified.
+- `daily` command.
 #### Changed
 - Using `giveaway` command requires "manage guild" permission.
 - Increased `meme`'s cooldown from 4s to 6s.